+++ inst/doc/cstidy.R
@@ -10,14 +10,14 @@ library(magrittr)
 
 ## -----------------------------------------------------------------------------
 d <- cstidy::generate_test_data()
-cstidy::set_csfmt_rts_data_v1(d)
+cstidy::set_csfmt_rts_data_v2(d)
 
 # Looking at the dataset
 d[]
 
 ## -----------------------------------------------------------------------------
 d <- cstidy::generate_test_data()[1:5]
-cstidy::set_csfmt_rts_data_v1(d)
+cstidy::set_csfmt_rts_data_v2(d)
 
 # Looking at the dataset
 d[]
@@ -45,10 +45,10 @@ d
 # Collapsing down to different levels, and healing the dataset 
 # (so that it can be worked on further with regards to real time surveillance)
 d[, .(deaths_n = sum(deaths_n), location_code = "norge"), keyby=.(granularity_time)] %>%
-  cstidy::set_csfmt_rts_data_v1(create_unified_columns = FALSE) %>%
+  cstidy::set_csfmt_rts_data_v2(create_unified_columns = FALSE) %>%
   print()
 
-# Collapsing to different levels, and removing the class csfmt_rts_data_v1 because
+# Collapsing to different levels, and removing the class csfmt_rts_data_v2 because
 # it is going to be used in new output/analyses
 d[, .(deaths_n = sum(deaths_n), location_code = "norge"), keyby=.(granularity_time)] %>%
   cstidy::remove_class_csfmt_rts_data() %>%
@@ -56,12 +56,12 @@ d[, .(deaths_n = sum(deaths_n), location_code = "norge"), keyby=.(granularity_ti
 
 ## -----------------------------------------------------------------------------
 cstidy::generate_test_data() %>%
-  cstidy::set_csfmt_rts_data_v1() %>%
+  cstidy::set_csfmt_rts_data_v2() %>%
   summary()
 
 ## -----------------------------------------------------------------------------
 cstidy::generate_test_data() %>%
-  cstidy::set_csfmt_rts_data_v1() %>%
+  cstidy::set_csfmt_rts_data_v2() %>%
   cstidy::identify_data_structure("deaths_n") %>%
   plot()