+++ NEWS.md
@@ -1,3 +1,15 @@
+# Version 2025.10.27
+
+- Updating csfmt_rts_data_v2 to be in line with the newest cstime version 2025.10.13
+
+# Version 2024.6.13
+
+- Inclusion of season in csfmt_rts_data_v2
+
+# Version 2023.12.28
+
+- Inclusion of isoquarter and isoyearquarter in csfmt_rts_data_v2
+
 # Version 2023.5.22
 
 - CRAN Submission.